backend/requirements.txt
fastapi
uvicorn
pydantic
fal-client
ell-ai
